+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 1.1.0 (2017-03-31)
+
+### Features
+
+* **client:** (*camera*) Open a direct session with camera
+* **client:** (*camera*) Close session
+* **client:** (*camera*) Check credentials
+* **client:** (*camera*) Send custom GET request
+* **client:** (*camera*) Send multiple GET requests in a single command
+
 # 1.0.0 (2017-03-25)
 
 ### Features